# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Routes, note sets and notes for the RecceLog application.",
        "title": "RecceLog API",
        "contact": {
            "name": "RecceLog"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/routes": {
            "get": {
                "description": "Returns lightweight summaries (no path, no notes) — backed by a single-table scan.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "List all routes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.RouteSummaryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            },
            "post": {
                "description": "Atomically inserts the route, its waypoints and the first note set (with its notes). Queries are run in a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Create a route with an initial note set",
                "parameters": [
                    {
                        "description": "Route + note set payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.CreateRouteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.RouteDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Malformed JSON or unknown fields",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "422": {
                        "description": "Validation failed (missing note_set, bad enum, etc.)",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
        "/v1/routes/range/{radius}": {
            "get": {
                "description": "Spatial search using PostGIS ST_DWithin + GIST index. Center is read from headers, radius is the path param (meters). Optionally filter by vehicle: pass one or more `vehicle` query params (e.g. ?vehicle=CAR\u0026vehicle=BIKE) to return only routes traversable by at least one of them.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "List routes within a radius",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Search radius in meters",
                        "name": "radius",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Filter: routes traversable by at least one of these vehicles",
                        "name": "vehicle",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Center latitude (WGS84)",
                        "name": "Latitude",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Center longitude (WGS84)",
                        "name": "Longitude",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.RouteSummaryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid radius or missing headers",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "422": {
                        "description": "Coordinates out of range, radius too large or invalid vehicle",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
        "/v1/routes/{id}": {
            "get": {
                "description": "Composes the route (with waypoints) and all its note sets (with notes) in two queries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Get a route with its note sets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.RouteDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid route id",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a route owned by the caller; cascades to its waypoints, note sets and notes. Only the route author may delete it.",
                "tags": [
                    "routes"
                ],
                "summary": "Delete a route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid route id",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "403": {
                        "description": "Not the route author",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
        "/v1/routes/{id}/notes": {
            "post": {
                "description": "Inserts a new note set (and its notes) for the given route in a single transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Append a note set to a route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Note set payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteSetInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteSetResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid route id or malformed JSON",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
        "/v1/routes/{id}/waypoints": {
            "post": {
                "description": "Inserts a single waypoint for the given route. Returns 404 if the route does not exist.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Add a waypoint to a route",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Waypoint payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.WaypointIn"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.WaypointResp"
                        }
                    },
                    "400": {
                        "description": "Invalid route id or malformed JSON",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "403": {
                        "description": "Not the route author",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_RecceLog_api_internal_http_dto.CreateRouteRequest": {
            "type": "object",
            "properties": {
                "note_set": {
                    "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteSetInput"
                },
                "route": {
                    "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.RouteInput"
                }
            }
        },
...
        "github_com_RecceLog_api_internal_http_dto.NoteSetInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteIn"
                    }
                }
            }
        },
        "github_com_RecceLog_api_internal_http_dto.NoteSetResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteResponse"
                    }
                },
                "route_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_RecceLog_api_internal_http_dto.RouteSummaryResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "cover_photo_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "finish_city": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "length_m": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "start_city": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "vehicles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_RecceLog_api_internal_http_problem.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "violations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Violation"
                    }
                }
            }
        },
        "github_com_RecceLog_api_internal_http_problem.Violation": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}